+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 0.1.0 (unreleased)
 ==================
 - **New resource:** `amplience_content_repository`
+- **New resource:** `amplience_content_webhook`
+- POC placeholder Amplience client abstraction package